requirements.txt
pyserial
pyinstaller
pycodestyle